Extract tests for isUint16Array to own file
